refactor: move page related components to containers folder
