+++ src/components/containers/About/Career.tsx
@@ -1,6 +1,6 @@
 import { format, parseISO } from 'date-fns';
-import { jobs } from '../../data/career';
-import { getDurationString } from '../../utils/getDurationString';
+import { jobs } from '../../../data/career';
+import { getDurationString } from '../../../utils/getDurationString';
 
 export interface Job {
   jobTitle: string;
