+++ src/components/containers/Contact/styles.ts
@@ -1,4 +1,4 @@
-import { styled } from '../../../stitches.config';
+import { styled } from '../../../../stitches.config';
 
 export const Form = styled('form', {
   display: 'flex',
