+++ src/components/containers/Projects/AllProjects.tsx
@@ -1,4 +1,4 @@
-import { projects } from '../../data/projects';
+import { projects } from '../../../data/projects';
 
 export const AllProjects = () => {
   return (
