+++ src/components/containers/Projects/styles.ts
@@ -1,5 +1,5 @@
 import { motion } from 'framer-motion';
-import { styled } from '../../../stitches.config';
+import { styled } from '../../../../stitches.config';
 
 export const StyledFeaturedProjects = styled(motion.div, {
   margin: '0.625rem 0 0 -1.25rem',
